test: refactor proposal flows
